requirements.txt
requests
pandas
python-dotenv
nltk
spacy
